// backend/migrations/snapshots/5dc8eba0327a0d53287bd1649382dcba066b078c1396ac64018c42f3603ec3a4/contract.d.ts
export type StorageHash =
  StorageHashBase<'5dc8eba0327a0d53287bd1649382dcba066b078c1396ac64018c42f3603ec3a4'>;
export type ExecutionHash =
  ExecutionHashBase<'e5be3e4360498dfa35fe67128e78083e7d0f24d196b23ca568273171b2c5c28c'>;
export type ProfileHash =
  ProfileHashBase<'3916f444a8a17ad749191acf9e08dad97d1a327b88c2f1d45d12f240296aa8b2'>;

export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Proyecto: { readonly idProyecto: CodecTypes['pg/uuid@1']['output'] };
  };
};
export type FieldInputTypes = {
  readonly public: { readonly Proyecto: { readonly idProyecto: CodecTypes['pg/uuid@1']['input'] } };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly gp_proyecto: { readonly idProyecto: CodecTypes['pg/uuid@1']['output'] };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly gp_proyecto: { readonly idProyecto: CodecTypes['pg/uuid@1']['input'] };
  };
};

export namespace Models {
  export type public_Proyecto = {
    idProyecto: CodecTypes['pg/uuid@1']['output'];
    readonly [RelationKeys]?: never;
  };
}

export declare const models: {
  public: {
    Proyecto: Models.public_Proyecto;
  };
};

export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;

type ContractBase = Omit<
  ContractType<{
    readonly namespaces: {
      readonly public: {
        readonly id: 'public';
        readonly kind: 'postgres-schema';
        readonly entries: {
          readonly table: {
            readonly gp_proyecto: {
              columns: {
                readonly idProyecto: {
                  readonly nativeType: 'uuid';
                  readonly codecId: 'pg/uuid@1';
                  readonly nullable: false;
                };
              };
              primaryKey: { readonly columns: readonly ['idProyecto'] };
              uniques: readonly [];
              indexes: readonly [];
              foreignKeys: readonly [];
            };
          };
        };
      };
    };
    readonly storageHash: StorageHash;
  }>,
  'roots' | 'domain'
> & {
  readonly target: 'postgres';
  readonly targetFamily: 'sql';
  readonly roots: {
    readonly gp_proyecto: {
      readonly namespace: 'public' & NamespaceId;
      readonly model: 'Proyecto';
    };
  };
  readonly domain: {
    readonly namespaces: {
      readonly public: {
        readonly models: {
          readonly Proyecto: {
            readonly fields: {
              readonly idProyecto: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/uuid@1' };
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: {
              readonly table: 'gp_proyecto';
              readonly namespaceId: 'public';
              readonly fields: { readonly idProyecto: { readonly column: 'idProyecto' } };
            };
          };
        };
      };
    };
  };
  readonly capabilities: {
    readonly postgres: {
      readonly distinctOn: true;
      readonly jsonAgg: true;
      readonly lateral: true;
      readonly limit: true;
      readonly orderBy: true;
      readonly returning: true;
    };
    readonly sql: {
      readonly checkConstraint: true;
      readonly defaultInInsert: true;
      readonly enums: true;
      readonly lateral: true;
      readonly returning: true;
      readonly scalarList: true;
    };
  };
  readonly extensions: {};
  readonly execution: {
    readonly executionHash: ExecutionHash;
    readonly mutations: {
      readonly defaults: readonly [
        {
          readonly ref: {
            readonly namespace: 'public';
            readonly table: 'gp_proyecto';
            readonly column: 'idProyecto';
          };
          readonly onCreate: { readonly kind: 'generator'; readonly id: 'uuidv4' };
        },
      ];
    };
  };
  readonly meta: {};

  readonly profileHash: ProfileHash;
};

export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;